requirements.txt
Babel==1.3
argparse==1.2.1
certifi==14.05.14
iso8601==0.1.10
pbr==0.10.0
prettytable==0.7.2
python-novaclient>=2.17.0
pytz==2014.7
requests==2.4.0
simplejson==3.6.3
six==1.7.3
wsgiref==0.1.2
